% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{assign_cbsa}
\alias{assign_cbsa}
\title{Assign CBSA}
\usage{
assign_cbsa(tract, county, date = format(Sys.Date(), "\%Y\%m"),
  use_md = TRUE)
}
\arguments{
\item{tract}{Census tract}

\item{county}{5-digit county FIPS code}

\item{date}{Date of OMB defintition to use (default equals current definition)}

\item{only_metro}{Match only Metropolitan Statistical Areas (default = FALSE)}
}
\description{
This function assigns the CBSA for a supplied county or Census tract. The
assignment will be based on the CBSA definitions that are in place on January
1st of each date by default (assuming only the year is supplied).
}
