% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{assign_lmi}
\alias{assign_lmi}
\title{Assign Income Level}
\usage{
assign_lmi(tract, year, return_label = TRUE)
}
\arguments{
\item{tract}{Numeric vector of Census tract codes (11 digits)}

\item{year}{A numeric year (YYYY)}

\item{return_label}{A logical indicating if the income level categorization is
to be returned or the relative income (default = TRUE)}
}
\value{
Either the relative income (if return_label == FALSE) or the income level
}
\description{
This function takes a vector of Census tracts and a year and returns either
the relative income of the tract or the income level categorization (i.e., low,
moderate, middle, or upper).
}
