% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{assign_mfi}
\alias{assign_mfi}
\title{Assign Median Family Income}
\usage{
assign_mfi(cbsa, year, use_hud = FALSE)
}
\arguments{
\item{cbsa}{Vector of CBSAs}

\item{year}{Calendar year of MFI data to use}

\item{use_hud}{Logical indicating whether HUD estimates are to be used or
Census values (default = FALSE)}
}
\description{
Assigns the appropriate median family income (based on census or HUD estimates).
}
