% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importing_data.R
\name{convert2numeric}
\alias{convert2numeric}
\title{Convert to Numeric}
\usage{
convert2numeric(df)
}
\description{
This function will convert all of the numeric variables from character into
numeric.  The reason I cannot just specify numeric formats for these variables
in the read_xls function is that the input files have notes at the bottom that
are read into some of the numeric variables, resulting in an error.
}
