% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{determine_file_date}
\alias{determine_file_date}
\title{Determine File Date}
\usage{
determine_file_date(date)
}
\arguments{
\item{date}{A numeric date of the format YYYYMM}
}
\description{
This function determines the CBSA definitions that were in effect during the
date specified. This is an internal function.
}
