% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{download_and_save_shapefile}
\alias{download_and_save_shapefile}
\title{Download and Save Shapefile}
\usage{
download_and_save_shapefile(basename)
}
\arguments{
\item{basename}{Character scalar giving the base name of the geographic file
to read.}
}
\value{
A `SpatialPolygonsDataFrame` (from the `sp` package) that includes the
contents of the TIGER line file converted to a format usable in R.
}
\description{
If the RDS file is not found for the required geography, this function downloads
the required file from the Census website and creates the necessary file.
}
