% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importing_data.R
\name{get_necta_specs}
\alias{get_necta_specs}
\title{Returns the specifications of the NECTA definition file}
\usage{
get_necta_specs(date_stamp)
}
\description{
This is an internal function used by import_omb.
}
