% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{get_state_ind}
\alias{get_state_ind}
\title{Get State Indicator}
\usage{
get_state_ind(shapes, state, geo)
}
\arguments{
\item{shapes}{A SpatialPolygonsDataFrame}

\item{state}{Character scalar giving the name of the geography to identify (e.g.,
alaska, hawaii, excluded, teritories)}

\item{geo}{Character vector giving the geography of the shape file (e.g, state,
cbsa)}
}
\value{
Boolean vector indicating which rows correspond to the states specified
in the function call.
}
\description{
Determines which rows of a SpatialPolygonsDataFrame corresponds to the states
specified in the function call.
}
