% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importing_data.R
\name{import_census}
\alias{import_census}
\title{Import Census Data}
\usage{
import_census()
}
\description{
This function uses the tidycensus package to import data on median family incomes
at the Census tract level from either the 2000 Decennial Census or the appropriate
5-year American Community Survey.
}
