% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importing_data.R
\name{import_distressed}
\alias{import_distressed}
\title{Import Distressed Areas}
\usage{
import_distressed(year)
}
\arguments{
\item{year}{4-digit year of file to be read}
}
\description{
Reads in information on distressed or underserved tracts.
}
