% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importing_data.R
\name{import_ffiec}
\alias{import_ffiec}
\title{Import FFIEC Median Family Incomes}
\usage{
import_ffiec(year)
}
\arguments{
\item{year}{4-digit year of file to be read in}
}
\description{
This function loans the median family incomes for MSAs and for the nonmetropolitan
state areas necessary to determine the relative income level of a Census tract.
These figures are published each year by the Federal Financial Institutions
Examination Council (FFIEC).
}
