% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importing_data.R
\name{import_mfi}
\alias{import_mfi}
\title{Import FFIEC Median Family Incomes}
\usage{
import_mfi(year)
}
\arguments{
\item{year}{4-digit year of file to be read in}
}
\description{
This function loans the median family incomes for MSAs and for the nonmetropolitan
state areas necessary to determine the relative income level of a Census tract.
}
