% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importing_data.R
\name{import_omb}
\alias{import_omb}
\title{Import OMB CBSA Definitions}
\usage{
import_omb(in_file)
}
\arguments{
\item{in_file}{Location of file to import. The file name must start with
'cbsa' or 'necta', be an Excel file (xls), and have a date string embedded
into the title as YYYYMM.}
}
\description{
This function will import MSA definitons as published by OMB.
}
