% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_functions.R
\name{list_state_fips}
\alias{list_state_fips}
\title{List of State FIPs Codes}
\usage{
list_state_fips(use_territories = TRUE)
}
\description{
This function returns a list of available state FIPs codes in the same order
as the state abbreviations and names.
}
