% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_functions.R
\name{list_state_names}
\alias{list_state_names}
\title{List of State Names}
\usage{
list_state_names(use_territories = TRUE)
}
\arguments{
\item{use_territories}{Should US territories be included with state names (default = TRUE)}
}
\description{
Returns the state names.  Supplements the default variable state.name to include
the District of Columbia and the U.S. territories.
}
