% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{load_cbsa}
\alias{load_cbsa}
\title{Load CBSA Data.frame}
\usage{
load_cbsa(date)
}
\arguments{
\item{date}{Date for the file to be returned.  Can be a year or a year-month.}
}
\description{
This function will load the appropriate CBSA data.frame for a given date.
}
