% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{load_census}
\alias{load_census}
\title{Load Census Data}
\usage{
load_census()
}
\description{
Loads an existing Census data file created by import_census.  The file contains
a data.frame with 3 columns:  tract, tract_mfi, and source file.
}
