% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{load_necta}
\alias{load_necta}
\title{Load NECTA Data.frame}
\usage{
load_necta(date)
}
\arguments{
\item{date}{Date for the file to be returned. Can be a year or a year-month.}
}
\description{
This function will load the appropriate NECTA data.frame for a given date.
}
