% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importing_data.R
\name{process_necta}
\alias{process_necta}
\title{Process NECTA file}
\usage{
process_necta(in_file)
}
\arguments{
\item{in_file}{Location of file to import. The file name must start with
'cbsa' or 'necta', be an Excel file (xls), and have a date string embedded
into the title as YYYYMM.}
}
\description{
This internal function reads in the NECTA definition file from OMB.
}
