% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{relinc2lmi}
\alias{relinc2lmi}
\title{Convert Relative Income to Income Category}
\usage{
relinc2lmi(x)
}
\arguments{
\item{x}{A numericn relative incomne}
}
\description{
Converts a relative income into a category that indicates if it is considered
low, moderate, middle, or upper income.
}
