% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{source_tiger_map_years}
\alias{source_tiger_map_years}
\title{Source TIGER Map Years}
\usage{
source_tiger_map_years()
}
\value{
Integer vector of years for which TIGER line data are available
}
\description{
Looks up the range of years for which TIGER map data are available
online.
}
