% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importing_data.R
\name{state2fips}
\alias{state2fips}
\title{Convert State to State FIPS}
\usage{
state2fips(s, use_name = FALSE)
}
\arguments{
\item{s}{State as a character vector. Can be state name or abbreviation}

\item{use_name}{Input variable is the name of the staes (default = FALSE)}
}
\description{
This function takes a state (either the name of the state or the two-character
abbreviation) and returns the numeric state FIPS.
}
