% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importing_data.R
\name{state_abbs}
\alias{state_abbs}
\title{List of State Abbreviations}
\usage{
state_abbs(use_territories = TRUE)
}
\description{
Supplements the options in state.abb to include DC and the U.S. Territories.
}
