% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importing_data.R
\name{state_fips}
\alias{state_fips}
\title{List of State FIPs Codes}
\usage{
state_fips()
}
\description{
This function returns a list of available state FIPs codes in the same order
as the state abbreviations and names.
}
