% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importing_data.R
\name{state_names}
\alias{state_names}
\title{List of State Names}
\usage{
state_names()
}
\description{
Returns the state names.  Supplements the default variable state.name to include
the District of Columbia and the U.S. territories.
}
